Fixed up throwing so it now using a projectile weapon and added ranged weapons that use the same thing.  Projectiles are physics based real objects.

+++ Assets/Scripts/Items/HealthPickup.cs
@@ -5,7 +5,7 @@ using MedievalMayhem.Entites;
 namespace MedievalMayhem.Items {
 	public class HealthPickup : ItemPickup
 	{
-		public int healingCapacity = 10;
+		public float healingCapacity = 10;
 
 		protected override void Start() {
 			this.SetPickupType (ItemPickup.POWERUP);
